+++ docnexus/version_info.py
@@ -1,3 +1,3 @@
 __version__ = '1.2.5'
-__build_timestamp__ = '2026-01-11 18:45:51'
+__build_timestamp__ = '2026-01-11 19:42:10'
 __build_type__ = 'Dev'